fix(Pointer): ensure direction indicator is only set on touch
The Pointer Direction Indicator would always have its rotation set
in the update even when releasing the touchpad which would cause
the pointer direction indicator rotation to be reset to the default
and make it look as if it snapped back to the original rotation.

This fix only updates the rotation if the touchpad is being touched.